requirements.txt
httpx>=0.27
beautifulsoup4>=4.12
fastapi>=0.110
uvicorn>=0.27
jinja2>=3.1
pyyaml>=6.0
rich>=13.0
python-dotenv>=1.0
apscheduler>=3.10
pyjwt>=2.8
bcrypt>=4.1
pyotp>=2.9
qrcode>=7.4
aiosqlite>=0.19
